Allow creating swap filesystems
The Ansible filesystem module does not support resizing swap
filesystems, and gives the following error when resizefs is true (which
was previously true by default):

  module does not support resizing swap filesystem yet.

In PR#89 this was worked around by skipping filesystem creation for swap
volumes. This isn't ideal however, because it prevents creation of
swapfs filesystems using this role, and will silently ignore requests to
create them.

This change fixes the issue by setting resizefs to false when
filesystem=swap.

Fixes: #88